move proxy file to bokeh dir and fix imports
